+++ LibraryNotes/DocumentList/DocumentListViewController.swift
@@ -483,6 +483,17 @@ extension MFMailComposeResult: CustomStringConvertible {
 // MARK: - DocumentTableControllerDelegate
 
 extension DocumentListViewController: DocumentTableControllerDelegate {
+  func documentTableControllerDidScroll(_ documentTableController: DocumentTableController) {
+    guard
+      let searchController = navigationItem.searchController,
+      let text = searchController.searchBar.text,
+      !text.isEmpty
+    else {
+      return
+    }
+    searchController.searchBar.resignFirstResponder()
+  }
+
   func showQuotes(quotes: [ContentIdentifier], shiftFocus: Bool) {
     let quotesVC = QuotesViewController(database: database)
     quotesVC.quoteIdentifiers = quotes